API completa!

+++ ecommerce/src/main/java/br/org/serratec/ecommerce/services/CategoriaService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.org.serratec.ecommerce.entities.Categoria;
+import br.org.serratec.ecommerce.exception.NoSuchElementException;
+import br.org.serratec.ecommerce.exception.PropertyValueException;
 import br.org.serratec.ecommerce.repositories.CategoriaRepository;
 
 @Service
@@ -18,42 +20,42 @@ public class CategoriaService {
 		return categoriaRepository.findAll();
 	}
 
-	public Categoria findById(Integer id) {
-		return categoriaRepository.findById(id).get();
+	public Categoria findById(Long id) {
+		return categoriaRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Categoria", id));
 	}
 
 	public Categoria save(Categoria categoria) {
+		if (categoria.getNome() == null || categoria.getDescricao() == null) {
+			throw new PropertyValueException("Categoria");
+		}
 		return categoriaRepository.save(categoria);
 	}
 
 	public Categoria update(Categoria categoria) {
+		if (categoria.getNome() == null || categoria.getDescricao() == null) {
+			throw new PropertyValueException("Categoria");
+		}
 		return categoriaRepository.save(categoria);
 	}
 
-	public void delete(Categoria categoria) {
-		categoriaRepository.delete(categoria);
-	}
-
-	public void deleteById(Integer id) {
-		categoriaRepository.deleteById(id);
-	}
-
-	public long count() {
-		return categoriaRepository.count();
-	}
-
-	public boolean deleteById2(Integer id) {
-		if (categoriaRepository.existsById(id)) {
-			categoriaRepository.deleteById(id);
-			Categoria categoriaDeletado = categoriaRepository.findById(id).orElse(null);
-			if (categoriaDeletado == null) {
-				return true;
-			} else {
-				return false;
-			}
-		} else {
+	public boolean delete(Categoria categoria) {
+		// verifica se categoria é valido
+		if (categoria == null)
 			return false;
-		}
+
+		// verifica se existe no banco
+		Categoria categoriaExistente = findById(categoria.getIdCategoria());
+		if (categoriaExistente == null)
+			return false;
+
+		// deleta categoria
+		categoriaRepository.delete(categoria);
+
+		// verifica se foi deletado de fato
+		Categoria categoriaContinuaExistindo = findById(categoria.getIdCategoria());
+		if (categoriaContinuaExistindo == null)
+			return true;
+		return false;
 	}
 
 }
